Separated project to Main.py and Table.py files
Main.py will be the main execution of program And table.py will be actually making all the tables with the tags, and inserting data into them. 

A text file should be added so that more videos can fill up the data base. And search function needs to be implemented as well.