Files via Upload
Progress being made. Now able to iterate through the tags dictionary and make tables on all the tags and inputting the title and url of those tags in that table. 

work to be done: search function implementation, opening urls and making sure that if table() is being run again that the same information is not being added to the tables unncessarily.